moved layout into components directory
