+++ client/src/Routes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Routes as Switch } from 'react-router-dom';
 
-import Layout from 'src/pages/layout';
+import Layout from 'src/components/layout';
 import Dashboard from 'src/pages/dashboard';
 import Login from 'src/pages/user/auth/login';
 import Register from 'src/pages/user/auth/register';
